Initial commit of the midi/ directory under audio containing the midi package.
The midi package will provide utilities for reading and writing MIDI files. This
first commit adds a midi-structs.go file that contains data structures and
constants that are used to represent a MIDI file in memory. It also adds midi.go
which contains logic for reading and writing MIDI files. Currently, only parsing
of the MIDI header is supported.